tradier-mcp/requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
mcp>=1.0.0
asyncio>=3.4.3
aiohttp>=3.8.0
pydantic>=2.0.0
